backend/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.4
distlib==0.3.2
Django==3.2.7
django-cors-headers==3.8.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
filelock==3.0.12
html5lib==1.1
idna==3.2
numpy==1.21.2
pandas==1.3.2
platformdirs==2.3.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.15.0
sqlparse==0.4.1
twstock==1.3.1
urllib3==1.26.6
virtualenv==20.7.2
webencodings==0.5.1
